Fix all fast refresh issues
