+++ src/components/Collection/constants.ts
@@ -1,0 +1,12 @@
+export const posterItemSize = {
+  width: 209,
+  height: 363,
+  gap: 16,
+};
+
+export const listItemSize = {
+  width: 907,
+  height: 319,
+  widthAlt: 907,
+  gap: 32,
+};
