+++ src/components/Dialogs/constants.ts
@@ -1,0 +1,52 @@
+export const languageDescription = {
+  'x-jat': 'Romaji (x-jat)',
+  'x-zht': 'Pinyin (x-zht)',
+  en: 'English (en)',
+  ja: 'Japanese (ja)',
+  ar: 'Arabic (ar)',
+  bd: 'Bangladeshi (bd)',
+  bg: 'Bulgarian (bd)',
+  ca: 'Canadian-French (ca)',
+  zh: 'Chinese',
+  'zh-hans': 'Chinese (zh-hans)',
+  'zh-hant': 'Chinese (zh-hant)',
+  cs: 'Czech (cs)',
+  cz: 'Czech (cz)',
+  da: 'Danish (da)',
+  dk: 'Danish (dk)',
+  nl: 'Dutch (nl)',
+  de: 'German (de)',
+  el: 'Greek (el)',
+  et: 'Estonian (et)',
+  fi: 'Finnish (fi)',
+  fr: 'French (fr)',
+  gl: 'Galician (gl)',
+  gr: 'Greek (gr)',
+  he: 'Hebrew (he)',
+  hu: 'Hungarian (hu)',
+  il: 'Hebrew (il)',
+  it: 'Italian (it)',
+  ko: 'Korean (ko)',
+  lv: 'Latvian (lv)',
+  lt: 'Lithuanian (lt)',
+  ms: 'Malaysian (ms)',
+  my: 'Malaysian (my)',
+  mn: 'Mongolian (mn)',
+  no: 'Norwegian (no)',
+  pl: 'Polish (pl)',
+  pt: 'Portuguese (pt)',
+  'pt-br': 'Portuguese - Brazil (pt-br)',
+  ro: 'Romanian (ro)',
+  ru: 'Russian (ru)',
+  sr: 'Serbian (sr)',
+  sk: 'Slovak (sk)',
+  sl: 'Slovenian (sl)',
+  es: 'Spanish (es)',
+  sv: 'Swedish (sv)',
+  se: 'Swedish (se)',
+  th: 'Thai (th)',
+  tr: 'Turkish (tr)',
+  uk: 'Ukrainian (uk)',
+  ua: 'Ukrainian (ua)',
+  vi: 'Vietnamese (vi)',
+};
