+++ src/pages/firstrun/AniDBAccount.tsx
@@ -6,8 +6,8 @@ import Input from '@/components/Input/Input';
 import TransitionDiv from '@/components/TransitionDiv';
 import { useAniDBTestLoginMutation } from '@/core/react-query/settings/mutations';
 import { setSaved as setFirstRunSaved, unsetSaved as unsetFirstRunSaved } from '@/core/slices/firstrun';
+import useFirstRunSettingsContext from '@/hooks/UseFirstRunSettingsContext';
 
-import { useFirstRunSettingsContext } from './FirstRunPage';
 import Footer from './Footer';
 
 import type { TestStatusType } from '@/core/slices/firstrun';
