+++ src/pages/firstrun/MetadataSourcesTabs/MovieDBTab.tsx
@@ -4,7 +4,7 @@ import cx from 'classnames';
 import Checkbox from '@/components/Input/Checkbox';
 import InputSmall from '@/components/Input/InputSmall';
 import TransitionDiv from '@/components/TransitionDiv';
-import { useFirstRunSettingsContext } from '@/pages/firstrun/FirstRunPage';
+import useFirstRunSettingsContext from '@/hooks/UseFirstRunSettingsContext';
 
 function MovieDBTab() {
   const { newSettings, updateSetting } = useFirstRunSettingsContext();
