+++ src/pages/settings/SettingsPage.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useMemo, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { useMediaQuery } from 'react-responsive';
 import { Outlet } from 'react-router';
-import { NavLink, useLocation, useOutletContext } from 'react-router-dom';
+import { NavLink, useLocation } from 'react-router-dom';
 import { mdiInformationOutline, mdiLoading } from '@mdi/js';
 import { Icon } from '@mdi/react';
 import cx from 'classnames';
@@ -15,8 +15,6 @@ import { usePatchSettingsMutation } from '@/core/react-query/settings/mutations'
 import { useSettingsQuery } from '@/core/react-query/settings/queries';
 import { setItem as setMiscItem } from '@/core/slices/misc';
 
-import type { SettingsType } from '@/core/types/api/settings';
-
 const items = [
   { name: 'General', path: 'general' },
   { name: 'Import', path: 'import' },
@@ -26,12 +24,6 @@ const items = [
   { name: 'User Management', path: 'user-management' },
   // { name: 'Themes', path: 'themes' },
 ];
-
-type ContextType = {
-  newSettings: SettingsType;
-  setNewSettings: (settings: SettingsType) => void;
-  updateSetting: (type: string, key: string, value: string | string[] | boolean) => void;
-};
 
 function SettingsPage() {
   const dispatch = useDispatch();
@@ -157,10 +149,6 @@ function SettingsPage() {
       />
     </div>
   );
-}
-
-export function useSettingsContext() {
-  return useOutletContext<ContextType>();
 }
 
 export default SettingsPage;
