+++ src/pages/utilities/UnrecognizedUtilityTabs/IgnoredFilesTab.tsx
@@ -12,6 +12,7 @@ import ItemCount from '@/components/Utilities/ItemCount';
 import MenuButton from '@/components/Utilities/Unrecognized/MenuButton';
 import Title from '@/components/Utilities/Unrecognized/Title';
 import UtilitiesTable from '@/components/Utilities/UtilitiesTable';
+import { staticColumns } from '@/components/Utilities/constants';
 import { useIgnoreFileMutation } from '@/core/react-query/file/mutations';
 import { useFilesInfiniteQuery } from '@/core/react-query/file/queries';
 import { useImportFoldersQuery } from '@/core/react-query/import-folder/queries';
@@ -20,9 +21,8 @@ import { FileSortCriteriaEnum, type FileType } from '@/core/types/api/file';
 import useEventCallback from '@/hooks/useEventCallback';
 import useFlattenListResult from '@/hooks/useFlattenListResult';
 import useRowSelection from '@/hooks/useRowSelection';
-import { staticColumns } from '@/pages/utilities/UnrecognizedUtility';
 
-import type { UtilityHeaderType } from '@/pages/utilities/UnrecognizedUtility';
+import type { UtilityHeaderType } from '@/components/Utilities/constants';
 import type { Updater } from 'use-immer';
 
 const Menu = (
